GRP-88-add-login #comment login working, need to fix login faled, group POJO created #time 5h
