--- src/main/java/io/github/bartmy/GroupS/LandingPage/login/LoginService.java
@@ -1,28 +1,0 @@
-package io.github.bartmy.GroupS.LandingPage.login;
-
-import io.github.bartmy.GroupS.user.UserRepository;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.stereotype.Service;
-
-
-@Slf4j
-@Service
-class LoginService {
-
-    private UserRepository repository;
-
-    LoginService(UserRepository repository){
-        this.repository = repository;
-    }
-
-    boolean loginValidation(String username, String password){
-        var loginOK = false;
-        if (repository.findByUsernameAndPassword(username, password).isEmpty()) {
-            log.info("login failed");
-        }else {
-            log.info("login success!");
-            loginOK = true;
-        }
-        return loginOK;
-    }
-}
