--- src/main/java/io/github/bartmy/GroupS/LandingPage/registration/Registration.java
@@ -1,37 +1,0 @@
-package io.github.bartmy.GroupS.LandingPage.registration;
-
-
-public class Registration {
-//
-//    public Registration(){
-//        registerNewUser();
-//    }
-//
-//    private void registerNewUser() {
-////        User user2 = new User("username", "password");
-//        User user = new User(1);
-//        user.registerUser();
-//        System.out.println("username from register: " + user.getUsername());
-//        if (usernameNotUsed(user.getUsername())) {
-//            System.out.println("you have been added!");
-//            addUserToDatabase(user.getUsername(), user.getPassword());
-//        } else {
-//            System.out.println("username taken, please select new");
-//        }
-//    }
-//    /**
-//     adding new user to mysql database
-//     */
-//    private void addUserToDatabase(String username, String password) {
-//        database.updateStatement(
-//                "INSERT INTO users(username, password) " +
-//                        "VALUES('" + username + "', '" + password + "');");
-//    }
-//    /**
-//     returning is username is available
-//     */
-//    private boolean usernameNotUsed(String username){
-//        return database.isUsernameTaken(username) == 0;
-//    }
-
-}
