+++ src/main/java/io/github/bartmy/GroupS/additionals/todo/Todo.java
@@ -1,0 +1,32 @@
+package io.github.bartmy.GroupS.additionals.todo;
+
+import lombok.Getter;
+import lombok.Setter;
+import org.hibernate.annotations.GenericGenerator;
+
+import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
+import javax.persistence.Id;
+import javax.persistence.Table;
+
+@Getter
+@Setter
+
+@Entity
+@Table(name = "todos")
+class Todo {
+
+    @Id
+    @GeneratedValue(generator="inc")
+    @GenericGenerator(name="inc", strategy = "increment")
+    private Integer id;
+    private String text;
+    private boolean done;
+
+    /**
+     * Hibernate (JPA) needs it.
+     */
+    @SuppressWarnings("unused")
+    Todo() {
+    }
+}
