+++ src/main/java/io/github/bartmy/GroupS/additionals/todo/TodoRepository.java
@@ -1,0 +1,8 @@
+package io.github.bartmy.GroupS.additionals.todo;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+interface TodoRepository extends JpaRepository<Todo, Integer> {
+}
