--- src/main/java/io/github/bartmy/GroupS/lang/Lang.java
@@ -1,36 +1,0 @@
-package io.github.bartmy.GroupS.lang;
-
-import lombok.Getter;
-import lombok.Setter;
-import org.hibernate.annotations.GenericGenerator;
-
-import javax.persistence.*;
-
-@Getter
-@Setter
-
-@Entity
-@Table(name = "languages")
-public class Lang {
-
-    @Id
-    @GeneratedValue(generator="inc")
-    @GenericGenerator(name="inc", strategy = "increment")
-    private Integer id;
-    @Column(name = "welcomemsg")
-    private String welcomeMsg;
-    private String code;
-
-    /**
-     * Hibernate (JPA) needs it.
-     */
-    @SuppressWarnings("unused")
-    Lang() {
-    }
-
-    public Lang(Integer id, String welcomeMsg, String code) {
-        this.id = id;
-        this.welcomeMsg = welcomeMsg;
-        this.code = code;
-    }
-}
