--- src/main/java/io/github/bartmy/GroupS/lang/LangRepository.java
@@ -1,11 +1,0 @@
-package io.github.bartmy.GroupS.lang;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-import java.util.List;
-
-@Repository
-public interface LangRepository extends JpaRepository<Lang, Integer> {
-    List<Lang> findByCodeAndWelcomeMsg(String code, String msg);
-}
