--- src/main/java/io/github/bartmy/GroupS/lang/LangService.java
@@ -1,26 +1,0 @@
-package io.github.bartmy.GroupS.lang;
-
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.stereotype.Service;
-
-import java.util.List;
-
-import static java.util.stream.Collectors.toList;
-
-@Slf4j
-@Service
-class LangService{
-    private LangRepository repository;
-
-    LangService(LangRepository repository){
-        this.repository = repository;
-    }
-
-    List<LangDTO> findAll(){
-        return repository
-                .findAll()
-                .stream()
-                .map(LangDTO::new)
-                .collect(toList());
-    }
-}
