--- src/main/java/io/github/bartmy/GroupS/todo/TodoRepository.java
@@ -1,8 +1,0 @@
-package io.github.bartmy.GroupS.todo;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-interface TodoRepository extends JpaRepository<Todo, Integer> {
-}
