+++ src/main/java/io/github/bartmy/GroupS/userProfile/groups/GroupDTO.java
@@ -1,0 +1,17 @@
+package io.github.bartmy.GroupS.userProfile.groups;
+
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@Setter
+class GroupDTO {
+    private Integer id;
+    private String groupName;
+
+    public GroupDTO(Group group){
+        this.id = group.getId();
+        this.groupName = group.getGroupName();
+    }
+
+}
